+++ ThirdPartyAdapters/applovin/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## AppLovin Android Mediation Adapter Changelog
 
+#### Next Version
+- Updated `minSdk` to 23
+- Rewarded ads now calls `onUserEarnedReward` without parameters.
+
 #### Version 13.1.0.0
 - Verified compatibility with AppLovin SDK 13.1.0.